Refactored tests to examples
